+++ src/dexterity/localroles/utils.py
@@ -1,15 +1,18 @@
 # -*- coding: utf-8 -*-
-from persistent.mapping import PersistentMapping
-from zope.component import getUtility
-from zope.component.interfaces import ComponentLookupError
-from zope.annotation.interfaces import IAnnotations
+from copy import deepcopy
 
-from Products.CMFCore.WorkflowCore import WorkflowException
-from Products.CMFPlone.utils import base_hasattr
+from dexterity.localroles.browser.settings import LocalRoleListUpdatedEvent
+from . import logger
+from persistent.mapping import PersistentMapping
 from plone import api
 from plone.dexterity.interfaces import IDexterityFTI
+from Products.CMFCore.WorkflowCore import WorkflowException
+from Products.CMFPlone.utils import base_hasattr
+from zope import event
+from zope.annotation.interfaces import IAnnotations
+from zope.component import getUtility
+from zope.component.interfaces import ComponentLookupError
 
-from . import logger
 
 rel_key = 'd.lr.related'
 
@@ -115,17 +118,22 @@ def add_fti_configuration(portal_type, configuration, keyname='static_config', f
         return "The portal type '%s' doesn't exist" % portal_type
     if not base_hasattr(fti, 'localroles'):
         setattr(fti, 'localroles', PersistentMapping())
+    old_value = fti.localroles.get(keyname, {})
     if keyname in fti.localroles and not force:
         logger.warn("The '%s' configuration on type '%s' is already set" % (keyname, portal_type))
         return "The '%s' configuration on type '%s' is already set" % (keyname, portal_type)
     fti.localroles[keyname] = configuration
+    event.notify(LocalRoleListUpdatedEvent(fti, keyname, old_value, configuration))
 
 
-def update_roles_in_fti(portal_type, configuration, action='add', keyname='static_config'):
+def update_roles_in_fti(portal_type, configuration, action='add', keyname='static_config', notify=True):
     """Update a localroles fti configuration by adding the given roles.
     :param portal_type: name of the portal type
     :param configuration: dict like {state: {principal: {'roles': [roles], 'rel': "{'utility name':[roles]}"}}}
+    :param action: select action to do (Default='add')
     :param keyname: 1st level key in localroles attr (Default='static_config')
+    :param notify: notify change to recatalog
+    :rtype: boolean, to indicate a change
     """
     # TODO : manage other actions
     try:
@@ -139,6 +147,7 @@ def update_roles_in_fti(portal_type, configuration, action='add', keyname='stati
     if keyname not in fti.localroles:
         fti.localroles[keyname] = {}
     lrd = fti.localroles[keyname]
+    old_value = deepcopy(lrd)
     for state in configuration:
         if state not in lrd:
             lrd[state] = {}
@@ -157,6 +166,21 @@ def update_roles_in_fti(portal_type, configuration, action='add', keyname='stati
                     change = True
     if change:
         fti.localroles._p_changed = True
+        if notify:
+            event.notify(LocalRoleListUpdatedEvent(fti, keyname, old_value, lrd))
+    return change
+
+
+def update_security_index(portal_types, trace=0):
+    """Update security index of portal_types"""
+    pc = api.portal.get_tool('portal_catalog')
+    with api.env.adopt_roles(['Manager']):
+        brains = pc.searchResults(portal_type=portal_types)
+        for i, brain in enumerate(brains, 1):
+            obj = brain.getObject()
+            if trace and i % trace == 0:
+                logger.info('On brain {}'.format(i))
+            obj.reindexObjectSecurity()
 
 
 def get_state(obj):